backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
# google-generativeai==0.3.2  # Removed - offline mode, no external API calls
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
aiofiles==23.2.1
# httpx==0.25.2  # Optional - only needed if external HTTP calls required
psutil==5.9.8
pynvml==11.5.0
python-evdev==1.9.0
pyserial==3.5

